backend/requirements.txt
Flask==2.3.6
mysql-connector-python==8.1.0